all task ready

+++ src/main/java/entity/store/Payment.java
@@ -18,7 +18,7 @@ public class Payment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "payment_id")
-    private Byte id;
+    private Short id;
 
     @ManyToOne
     @JoinColumn(name = "customer_id")
